dont use an extra interface
