--- src/main/java/com/publicuhc/pluginframework/routing/tester/DefaultPermissionTester.java
@@ -1,65 +1,0 @@
-/*
- * DefaultPermissionTester.java
- *
- * Copyright (c) 2014. Graham Howden <graham_howden1 at yahoo.co.uk>.
- *
- * This file is part of PluginFramework.
- *
- * PluginFramework is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * PluginFramework is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with PluginFramework.  If not, see <http ://www.gnu.org/licenses/>.
- */
-
-package com.publicuhc.pluginframework.routing.tester;
-
-import org.bukkit.ChatColor;
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandSender;
-
-import java.util.HashSet;
-
-/**
- * A tester that allows the command based on the permissions of the sender
- */
-public class DefaultPermissionTester extends HashSet<String> implements PermissionTester
-{
-    boolean matchingAll = false;
-
-    @Override
-    public boolean isMatchingAll()
-    {
-        return matchingAll;
-    }
-
-    @Override
-    public void setMatchingAll(boolean matching)
-    {
-        matchingAll = matching;
-    }
-
-    @Override
-    public boolean testCommand(Command command, CommandSender sender, String[] args)
-    {
-        //check permissions
-        for(String permission : this) {
-            if(sender.hasPermission(permission)) {
-                if(!matchingAll) {
-                    break;
-                }
-            } else {
-                sender.sendMessage(ChatColor.RED + "You do not have permission to run this command. (" + permission + ")");
-                return false;
-            }
-        }
-        return true;
-    }
-}
